+++ srcs/cub_3d/game/point_utils.c
@@ -6,7 +6,7 @@
 /*   By: rertzer <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/10 11:28:30 by rertzer           #+#    #+#             */
-/*   Updated: 2023/04/11 14:42:42 by rertzer          ###   ########.fr       */
+/*   Updated: 2023/04/12 15:33:54 by rertzer          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,10 +17,10 @@ bool	point_outofrange(t_cbdata *data, t_point *p)
 	bool	out;
 
 	out = false;
-	if (p->y >= (double)(data->map_y * BLOCK_SIZE) \
+	if (p->y >= (float)(data->map_y * BLOCK_SIZE) \
 			|| p->y <= 0.0 \
 			|| p->x <= 0.0 \
-			|| p->x >= (double)(data->map_x * BLOCK_SIZE))
+			|| p->x >= (float)(data->map_x * BLOCK_SIZE))
 	{
 		out = true;
 		p->x = INFINITY;
